lights/requirements.txt
# pip install --force-reinstall -r requirements.txt
numpy==1.26.4
matplotlib==3.8.4
rpi_ws281x
adafruit-circuitpython-neopixel
adafruit-blinka
RPi.GPIO
